Move validation-utility functions to module
